requirements.txt
pyserial
readchar
